+++ apps/minimal-tester/App.tsx
@@ -22,6 +22,8 @@ import {
   Image as RNImage,
 } from 'react-native';
 
+import DevMenuExample from './DevMenuExample';
+
 function randomColor() {
   return '#' + ((Math.random() * 0xffffff) << 0).toString(16).padStart(6, '0');
 }
@@ -50,6 +52,7 @@ export default function App() {
           isFabricEnabled: {isFabricEnabled + ''}
         </Text>
 
+        <DevMenuExample />
         <ImageExample />
         <LinearGradientExample />
         {Platform.OS === 'ios' && <BlurExample />}